refactor(api): replace /api/organizations/{organizationId}/campaigns accetptance tests by more complete ones in prescription/campaign
