--- api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -1,215 +1,0 @@
-import lodash from 'lodash';
-
-import { createServer } from '../../../../server.js';
-import { expect } from '../../../test-helper.js';
-import { databaseBuilder } from '../../../tooling/databases.js';
-import { generateAuthenticatedUserRequestHeaders } from '../../../tooling/test-utils/http-server.js';
-
-const { map: _map } = lodash;
-
-describe('Acceptance | Application | organization-controller', function () {
-  let server;
-
-  beforeEach(async function () {
-    server = await createServer();
-  });
-
-  describe('GET /api/organizations/{id}/campaigns', function () {
-    let campaignsData;
-    let organizationId, otherOrganizationId;
-    let userId;
-    let options;
-
-    beforeEach(async function () {
-      userId = databaseBuilder.factory.buildUser({}).id;
-      organizationId = databaseBuilder.factory.buildOrganization({}).id;
-      databaseBuilder.factory.buildMembership({ organizationId, userId });
-      otherOrganizationId = databaseBuilder.factory.buildOrganization({}).id;
-      campaignsData = _map(
-        [
-          { name: 'Quand Peigne numba one', code: 'ATDGRK343', organizationId },
-          { name: 'Quand Peigne numba two', code: 'KFCTSU984', organizationId },
-          {
-            name: 'Quand Peigne numba three',
-            code: 'ABC180ELO',
-            organizationId,
-            archivedAt: new Date('2000-01-01T10:00:00Z'),
-          },
-          {
-            name: 'Quand Peigne numba four',
-            code: 'ABC180LEO',
-            organizationId,
-            archivedAt: new Date('2000-02-01T10:00:00Z'),
-          },
-          { name: 'Quand Peigne otha orga', code: 'CPFTQX735', organizationId: otherOrganizationId },
-        ],
-        (camp) => {
-          const builtCampaign = databaseBuilder.factory.buildCampaign(camp);
-          return { name: camp.name, code: camp.code, id: builtCampaign.id };
-        },
-      );
-      databaseBuilder.factory.buildCampaignParticipation({ campaignId: campaignsData[4].id });
-      await databaseBuilder.commit();
-
-      options = {
-        method: 'GET',
-        url: `/api/organizations/${organizationId}/campaigns`,
-        headers: generateAuthenticatedUserRequestHeaders({ userId }),
-      };
-    });
-
-    context('Something is wrong', function () {
-      it('should respond with a 401 - unauthorized access - if user is not authenticated', function () {
-        // given
-        options.headers.authorization = 'invalid.access.token';
-
-        // when
-        const promise = server.inject(options);
-
-        // then
-        return promise.then((response) => {
-          expect(response.statusCode).to.equal(401);
-        });
-      });
-    });
-
-    context('when the user is not a member of the organization', function () {
-      it('should respond with a 403', function () {
-        // given
-        userId = databaseBuilder.factory.buildUser({}).id;
-        options = {
-          method: 'GET',
-          url: `/api/organizations/${organizationId}/campaigns`,
-          headers: generateAuthenticatedUserRequestHeaders({ userId }),
-        };
-
-        // when
-        const promise = server.inject(options);
-
-        // then
-        return promise.then((response) => {
-          expect(response.statusCode).to.equal(403);
-        });
-      });
-    });
-
-    context('Retrieve campaigns', function () {
-      it('should return 200 status code the organization campaigns', async function () {
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        const campaigns = response.result.data;
-        expect(campaigns).to.have.lengthOf(2);
-        expect(_map(campaigns, 'attributes.name')).to.have.members([campaignsData[0].name, campaignsData[1].name]);
-        expect(_map(campaigns, 'attributes.code')).to.have.members([campaignsData[0].code, campaignsData[1].code]);
-      });
-
-      it('should return campaigns with its owner', async function () {
-        // given
-        organizationId = databaseBuilder.factory.buildOrganization({}).id;
-        const ownerId = databaseBuilder.factory.buildUser({ firstName: 'Daenerys', lastName: 'Targaryen' }).id;
-        databaseBuilder.factory.buildMembership({ organizationId, userId });
-        databaseBuilder.factory.buildCampaign({ organizationId, ownerId });
-        await databaseBuilder.commit();
-        options.url = `/api/organizations/${organizationId}/campaigns`;
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.data[0].attributes['owner-first-name']).to.equal('Daenerys');
-        expect(response.result.data[0].attributes['owner-last-name']).to.equal('Targaryen');
-      });
-
-      it('should return archived campaigns', async function () {
-        // given
-        options.url = `/api/organizations/${organizationId}/campaigns?page[number]=1&page[size]=2&filter[status]=archived`;
-        const expectedMetaData = { page: 1, pageSize: 2, rowCount: 2, pageCount: 1, hasCampaigns: true };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-        expect(response.result.data).to.have.lengthOf(2);
-        expect(response.result.data[0].type).to.equal('campaigns');
-      });
-
-      it('should return report with the campaigns', async function () {
-        // given
-        databaseBuilder.factory.buildMembership({ organizationId: otherOrganizationId, userId });
-        await databaseBuilder.commit();
-        options.url = `/api/organizations/${otherOrganizationId}/campaigns`;
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        const campaigns = response.result.data;
-        expect(campaigns).to.have.lengthOf(1);
-        expect(response.result.data[0].attributes['participations-count']).to.equal(1);
-        expect(response.result.data[0].attributes['shared-participations-count']).to.equal(1);
-      });
-
-      it('should return default pagination meta data', async function () {
-        // given
-        const expectedMetaData = { page: 1, pageSize: 10, rowCount: 2, pageCount: 1, hasCampaigns: true };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-      });
-
-      it('should return a 200 status code with paginated data', async function () {
-        // given
-        options.url = `/api/organizations/${organizationId}/campaigns?&page[number]=2&page[size]=1`;
-        const expectedMetaData = { page: 2, pageSize: 1, rowCount: 2, pageCount: 2, hasCampaigns: true };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-        expect(response.result.data).to.have.lengthOf(1);
-        expect(response.result.data[0].type).to.equal('campaigns');
-      });
-
-      it('should return a 200 status code with paginated and filtered data', async function () {
-        // given
-        options.url = `/api/organizations/${organizationId}/campaigns?filter[name]=Two&page[number]=1&page[size]=1`;
-        const expectedMetaData = { page: 1, pageSize: 1, rowCount: 1, pageCount: 1, hasCampaigns: true };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-        expect(response.result.data).to.have.lengthOf(1);
-        expect(response.result.data[0].type).to.equal('campaigns');
-      });
-
-      it('should return a 200 status code with empty result', async function () {
-        // given
-        options.url = `/api/organizations/${organizationId}/campaigns?&page[number]=3&page[size]=1`;
-        const expectedMetaData = { page: 3, pageSize: 1, rowCount: 2, pageCount: 2, hasCampaigns: true };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-        expect(response.result.data).to.have.lengthOf(0);
-      });
-    });
-  });
-});
